give meteorological functions their own file
# Conflicts:
#	generic.py
